Modify ID datatype

+++ bl/src/main/java/com/techwizzle/site/repositories/CardRepository.java
@@ -3,6 +3,6 @@ package com.techwizzle.site.repositories;
 import com.techwizzle.site.model.Card;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CardRepository extends CrudRepository<Card, Integer> {
+public interface CardRepository extends CrudRepository<Card, Long> {
 
 }
